requirements.txt
streamlit
PyPDF2
langchain
langchain_community
openai
tiktoken
faiss-cpu
python-dotenv